+++ CHANGES.md
@@ -17,6 +17,7 @@ dev
 - Add `t.headset.hdr` and `lovr.headset.isHDR`.
 - Add `float16` and `int8` `GraphicsFeature`s.
 - Add `i8`, `i8x2`, `u8`, `u8x2`, `sn8`, `sn8x2`, `un8`, `un8x2`, `sn16`, `un16`, and `f16` `DataType`s.
+- Add `sleep` callback to `World:setCallbacks`.
 
 ### Change